Moved Login and Register views into account_forms/
